Make code coverage classes directory configurable
Description:

- WSO2 carbon products usually have the source code in plugins directory
- When it comes to vertical solutions like healthcare, BFSI - the source
code specific to the solutions goes under dropins folder of the base products.
- So, in the vertical solutions, when we run the integration tests, we need
the ability to point to dropins directory in order to generate the code
coverage reports of the solution.

Solution:

- Solution is to make this classes directory configurable via the automation.xml
- `coverageClassesRelativeDirectory` is an optional element under
automation/configurations element of the automation.xml

<!-- Set this to the relative path of the directory, where the classes you
need to instrument through code coverage tool, resides. By default this is
pointed to the plugins directory of a WSO2 Carbon Server -->
<coverageClassesRelativeDirectory>repository/components/dropins/
                                       </coverageClassesRelativeDirectory>

- Value of coverageClassesRelativeDirectory is a relative path from the
carbon_home.
- By default, when you don't set this in automation.xml, framework will
use the plugins directory as the classes directory for Jacoco report
generation.